#57 Refactored Setup action from BlogController into HomeController
