+++ src/Fan.Blogs/README.md
@@ -1,0 +1,8 @@
+﻿# Fan.Blogs
+
+## NuGet Packages
+
+Fan.Blogs tries to include everything related to the blog app (except views), BlogController, ViewModels in addition to Services and Data.  As a result, the following packages are required.
+
+- Microsoft.AspNetCore.Mvc.ViewFeatures which is Controller type resides
+- Microsoft.AspNetCore.Identity.EntityFrameworkCore is required by MetaWeblogService which uses UserManager and SignInManager
