+++ src/Fan.Web/Views/Home/Setup.cshtml
